Add middleware library functions.
These middleware include the error handling middleware and embeds
middleware. The embeds is responsible for initializing the
res.locals.embeds object so that controllers can easily push their
assets to this object.

The error handling middleware is responsible for errors that occur
within the application and displaying the error page to the user.